pyproject.toml
[project]
name = "image-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "scipy>=1.16.3",
]
